requirements.txt
amqp==5.1.1
asgiref==3.5.1
async-timeout==4.0.2
beautifulsoup4==4.11.1
billiard==3.6.4.0
bleach==5.0.0
celery==5.2.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0.4
django-celery-beat==2.2.1
django-celery-results==2.3.1
django-filter==21.1
django-post-office==3.6.0
django-timezone-field==4.2.3
djangorestframework==3.13.1
idna==3.3
jsonfield==3.1.0
kombu==5.2.4
lxml==4.8.0
Markdown==3.3.7
opentele==1.15.1
packaging==21.3
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
pyaes==1.6.1
pyasn1==0.4.8
pyparsing==3.0.8
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
PySocks==1.7.1
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
redis==4.2.2
requests==2.27.1
rsa==4.8
six==1.16.0
socks==0
soupsieve==2.3.2.post1
sqlparse==0.4.2
Telethon==1.24.0
TgCrypto==1.2.3
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.14.1
